Move Loader tests to LoaderIntegration.js and add new real unit test for InternalLoader.
